% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccs.R
\docType{class}
\name{CCS-class}
\alias{CCS-class}
\title{CCS-class}
\description{
CCS class
}
\section{Slots}{

\describe{
\item{\code{Repeat}}{The data for repeatability}

\item{\code{Data}}{Data about CCS probability and CCS subtypes}

\item{\code{Plot}}{a ggplot plot for CCS subtype visualization}
}}

\author{
Weibin Huang<\email{654751191@qq.com}>
}
\keyword{classes}
