% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccs.R
\name{completeModel,CCS-method}
\alias{completeModel,CCS-method}
\title{CCS method: completeModel}
\usage{
\S4method{completeModel}{CCS}(object, model.dir = NULL)
}
\arguments{
\item{model.dir}{a character. the path of model series.}
}
\value{
completeModel: a complete CCS class.
}
\description{
\code{completeModel} method for \code{CCS} class
}
