% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dr.R
\name{dr,CCS-method}
\alias{dr,CCS-method}
\title{CCS method: dr}
\usage{
\S4method{dr}{CCS}(
  object,
  method = c("UMAP", "t-SNE", "PCA")[1],
  dimension = c(2, 2),
  model.dir = NULL,
  seed = 456,
  numCores = 6,
  verbose = T,
  ...
)
}
\arguments{
\item{object}{a \code{\link{CCS-class}} object}

\item{method}{One of \code{'UMAP'} and \code{'t-SNE'}.}

\item{dimension}{the hyperparameters for 2 level t-SNE.}

\item{model.dir}{a character. the path of model series.}

\item{numCores}{No. of CPU core}

\item{verbose}{whether report messages}

\item{...}{Parameters of the core DR function.}
}
\value{
data frame. Series of PC1, PC2 and etc after dimensionality reduction
}
\description{
\code{dr} method for \code{CCS} class. Dimensionality reduction (DR) for CCS probability matrix.
}
\seealso{
\code{\link{ccs}}.
}
\author{
Weibin Huang<\email{hwb2012@qq.com}>
}
