% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{exploreCluster,CCS-method}
\alias{exploreCluster,CCS-method}
\title{CCS method: exploreCluster}
\usage{
\S4method{exploreCluster}{CCS}(
  object,
  nTop = 10,
  model.dir = NULL,
  size = 15,
  width = 12,
  height = 9,
  ...
)
}
\arguments{
\item{object}{a \code{\link{CCS-class}} object}

\item{nTop}{Integer. The count of top parameters you want to plot.}

\item{model.dir}{a character. the path of model series.}

\item{size}{The plot size}

\item{width}{The plot width}

\item{height}{The plot height}

\item{...}{Parameters in \code{\link{CCS-method.plot}}.}
}
\description{
\code{exploreCluster} method for \code{CCS} class
}
\seealso{
\code{\link{ccs}}; \code{\link{CCS-method.plot}}.
}
