% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select.R
\name{filterCCS,CCS-method}
\alias{filterCCS,CCS-method}
\title{CCS method: filterCCS}
\usage{
\S4method{filterCCS}{CCS}(
  object,
  model.dir = NULL,
  pattern = "blood",
  method = c("grep", "match")[1]
)
}
\arguments{
\item{object}{a \code{\link{CCS-class}} object}

\item{model.dir}{a character. the path of model series.}

\item{pattern}{if \code{method='match'}, it should be characters with the format \code{tissue-cohort}(such as \code{STAD-GSE22377}).}

\item{method}{One of \code{'grep'} or \code{match}.}
}
\value{
a new \code{\link{CCS-class}} object
}
\description{
\code{filterCCS} method for \code{CCS} class. filterCCS specified submodels
}
\seealso{
\code{\link{ccs}}.
}
\author{
Weibin Huang</email{hwb2012@qq.com}>
}
