% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{manualCluster,CCS-method}
\alias{manualCluster,CCS-method}
\title{CCS method: manualCluster}
\usage{
\S4method{manualCluster}{CCS}(object, classifier)
}
\arguments{
\item{object}{a \code{\link{CCS-class}} object}

\item{classifier}{a list of classifier.}
}
\description{
\code{manualCluster} method for \code{CCS} class
}
\seealso{
\code{\link{ccs}}.
}
