% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotBatchEffect.R
\name{plotBatchEffect,CCS-method}
\alias{plotBatchEffect,CCS-method}
\title{CCS method: plotBatchEffect}
\usage{
\S4method{plotBatchEffect}{CCS}(object, data)
}
\arguments{
\item{object}{a \code{\link{CCS-class}} object}
}
\description{
\code{plotBatchEffect} method for \code{CCS} class
}
\seealso{
\code{\link{ccs}}.
}
