% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccs.R
\name{plotCancerSubtype,CCS-method}
\alias{plotCancerSubtype,CCS-method}
\title{CCS method: plotCancerSubtype}
\usage{
\S4method{plotCancerSubtype}{CCS}(
  object,
  CCS = NULL,
  rm.zero = TRUE,
  add.cell_fun = TRUE,
  size = 15,
  verbose = TRUE
)
}
\arguments{
\item{object}{a \code{\link{CCS-class}} object}

\item{rm.zero}{Whether to remove zero value of CCS subtypes. In the strategy of \code{dbscan}, 0 means uncategorized, so you should set \code{rm.zero = TRUE}(default).}

\item{add.cell_fun}{Logic. Whether add a cell_fun into the heatmap.}

\item{size}{The plot size}

\item{verbose}{whether report messages}
}
\value{
plotCancerSubtype: a complete CCS class.
}
\description{
\code{plotCancerSubtype} method for \code{CCS} class. Plot CCS subtype across cancer types.
}
