% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccs.R
\name{predict.CCS}
\alias{predict.CCS}
\title{CCS method: predict}
\usage{
\method{predict}{CCS}(
  object,
  X,
  model.dir = NULL,
  work.space = paste0(model.dir, "/prediction/", project.name),
  cover = F,
  verbose = T,
  numCores = 4
)
}
\arguments{
\item{object}{a \code{\link{CCS-class}} object}

\item{X}{a numeric RNA expression vector or matrix with gene names}

\item{model.dir}{a character. the path of model series.}

\item{cover}{Whether to cover the existed result}

\item{verbose}{whether report messages}

\item{numCores}{No. of CPU core}

\item{project.name}{a project name about \code{X}. Because some might call lots of dataset based on one \code{CCS} model, It's recommended to set a projec name. Default is \code{01}.}
}
\value{
predict: A list object with CCS subtype prediction.
}
\description{
\code{predict} method for \code{CCS} class
}
\examples{
\dontrun{
ccs_pred <- predict(resCCS, model.dir = "./ccs/project_01")
}
}
\seealso{
\code{\link{ccs}}.
}
\author{
Weibin Huang<\email{hwb2012@qq.com}>
}
