% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccs.R
\name{CCS-method}
\alias{CCS-method}
\alias{predict.CCS}
\alias{plot.CCS}
\title{CCS-method}
\usage{
\method{predict}{CCS}(object, X, model.dir = NULL, verbose = T, numCores = 4)

\method{plot}{CCS}(object, size = 15)

\S4method{completeModel}{CCS}(object, model.dir = NULL)
}
\arguments{
\item{object}{a CCS object}

\item{X}{a numeric RNA expression vector or matrix with gene names}

\item{model.dir}{a character. the path of model series.}

\item{verbose}{whether report messages}

\item{numCores}{No. of CPU core}

\item{size}{the size of ggplot}
}
\value{
predict: A list object with CCS subtype prediction.

plot: A ggplot2 object

completeModel: a complete CCS class.
}
\description{
\code{predict} method for \code{CCS} class

\code{plot} method for \code{CCS} class

\code{completeModel} method for \code{CCS} class
}
\examples{
ccs_pred <- predict(resCCS, model.dir = "./ccs/project_01")
ccs_plot <- plot(resCCS)
}
\seealso{
\code{\link{ccs}}.

\code{\link{ccs}}.
}
\author{
Weibin Huang<\email{hwb2012@qq.com}>
}
