% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccs_base.R
\name{selectCCSData}
\alias{selectCCSData}
\title{selectCCSData}
\usage{
selectCCSData(
  dataList,
  geneSet,
  geneAnnotation = NULL,
  geneid = "ensembl",
  coreTissueType = c("ACC", "AEG", "BRCA", "CRC", "KIRC", "PAAD", "PRAD", "STAD",
    "Pediatric", "Blood"),
  minSampleSize = 20,
  maxMissingRate = 0.3,
  verbose = TRUE
)
}
\arguments{
\item{dataList}{A nested list object. The first layer is the type of tumor; the second layer is the gene expression matrix.}

\item{geneSet}{A list of genes for classification}

\item{coreTissueType}{Core tissue type you want for the CCS model establishment.}

\item{minSampleSize}{The lower margin of sample size of eligible cohorts.}

\item{maxMissingRate}{The upper margin of missing rate of eligible cohorts.}

\item{verbose}{whether report modeling process}
}
\value{
dataList for \code{\link{createCCSData}}.
}
\description{
Advanced selection of CCS data list
}
\seealso{
\code{\link{ccs}}; \code{\link{createCCSData}}.
}
\author{
Weibin Huang<\email{hwb2012@qq.com}>
}
