% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccs_base.R
\name{createCCSData}
\alias{createCCSData}
\title{createCCSData}
\usage{
createCCSData(dataList, GEPs, ...)
}
\arguments{
\item{dataList}{A nested list object. The first layer is the type of tumor; the second layer is the gene expression matrix.}

\item{GEPs}{List. Gene expression profiles - Co-expressed modules from WGCNA, etc}

\item{...}{parameters for \code{\link[GSClassifier]{subtypeVector}} function}
}
\value{
training data for \code{\link{ccs}} function
}
\description{
Create training data for CCS package based on expression matrix and GEPs
}
\seealso{
\code{\link{ccs}}.
}
\author{
Weibin Huang<\email{hwb2012@qq.com}>
}
