% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CCS_DX_mapping.R
\docType{data}
\name{CCS_DX_mapping}
\alias{CCS_DX_mapping}
\title{CCS Diagnosis Mapping}
\format{A data frame with 113440 obs. and 3 variables:
\describe{
  \item{category_code}{Numeric CCS category code}
  \item{code}{ICD-9 or ICD-10 diagnosis code, without prefixes or decimals}
  \item{code_type}{Either ICD-9-CM or ICD-10-CM}
}}
\source{
\url{https://www.hcup-us.ahrq.gov/toolssoftware/ccs/Multi_Level_CCS_2015.zip}

\url{https://www.hcup-us.ahrq.gov/toolssoftware/ccs10/ccs_dx_icd10cm_2019_1.zip}
}
\usage{
CCS_DX_mapping
}
\description{
A dataset containing the ICD-9-CM and ICD-10-CM codes for each CCS category
}
\seealso{
\code{\link{CCS_DX_categories}} for a list of CCS categories.
}
\keyword{datasets}
