% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{ccsCheck}
\alias{ccsCheck}
\title{Check whether CCS data list or submodels are well prepared}
\usage{
ccsCheck(
  data,
  geneSet,
  geneAnnotation,
  geneid = "ensembl",
  params = list(device = "cpu", nfold = 5, nrounds = 100, nthread = 2, eta = 0.3, gamma =
    0, max_depth = 14, colsample_bytree = 1, min_child_weight = 1, subsample = 0.7, n =
    30, sampSize = 0.7, ptail = 0.1, nround.mode = c("fixed", "polling")[1]),
  train.proc = 0.8,
  seed = 145,
  mode = c("dataset", "submodel")[1],
  nTest = 3,
  minAccuracy = 0.85,
  model.dir = "./ccs/PADv20240810",
  numCores = 16,
  verbose = TRUE
)
}
\arguments{
\item{data}{a list with components (expression matrix + subtype vector)}

\item{geneSet}{A list of genes for classification}

\item{geneAnnotation}{A data frame with ENSEMBL, SYMBOL and ENTREZID of the genes in \code{geneSet}}

\item{geneid}{type of gene id in \code{X}. One of \code{'symbol'}, \code{'entrez'} and \code{'ensembl'}.}

\item{params}{parameters of \code{\link[xgboost]{xgb.cv}}, \code{\link[xgboost]{xgboost}}, and \code{\link[GSClassifier]{fitEnsembleModel}}. Some important options are:
\itemize{
  \item \code{nfold} No. of cross validation subcohorts.
  \item \code{nrounds} Max number of boosting iterations. GSClassifier have optimized nrounds, so here I set a large value.
  \item \code{nthread} No. of CPU cores used in \code{\link[xgboost]{xgboost}}.
  \item \code{eta} Step size shrinkage used in update to prevent overfitting. Range = [0,1]
  \item \code{gamma} The larger gamma is, the more conservative the algorithm will be. Range = [0, Inf]
  \item \code{max_depth} Maximum depth of a tree. Deeper trees can capture more complex patterns in the data, but may also lead to overfitting. Range = [0, Inf]
  \item \code{colsample_bytree} Percentage of columns used for each tree construction. Lowering this value can prevent overfitting by training on a subset of the features. Range = (0, 1]
  \item \code{min_child_weight} The larger min_child_weight is, the more conservative the algorithm will be. Range = [0, Inf]
  \item \code{subsample} Preventing overfitting
  \item \code{n} \code{\link[GSClassifier]{fitEnsembleModel}} parameter
  \item \code{sampSize} \code{\link[GSClassifier]{fitEnsembleModel}} parameter. Range = (0,1]
  \item \code{ptail} \code{\link[GSClassifier]{fitEnsembleModel}} parameter. Range = (0,0.5]
}}

\item{mode}{One of \code{'dataset'} or \code{'submodel'}}

\item{nTest}{How many submodels you want to test}

\item{minAccuracy}{An acceptable accuracy}

\item{model.dir}{a character. the path of model series.}

\item{numCores}{No. of CPU core}

\item{verbose}{whether report messages}
}
\value{
data frame
}
\description{
Check whether CCS data list or submodels are well prepared
}
\seealso{
\code{\link{ccs}}.
}
\author{
Weibin Huang<\email{hwb2012@qq.com}>
}
