% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scaller.R
\docType{class}
\name{CCSDeepClassifier-class}
\alias{CCSDeepClassifier-class}
\title{S4 Class for a Custom Deep Learning Classifier}
\description{
This class encapsulates a trained deep learning model and the feature names required for prediction. It is designed to provide a standardized interface for deep learning models, particularly those trained with the 'luz' package.
}
\section{Slots}{

\describe{
\item{\code{model}}{A fitted model object. While typically a 'luz_module_fitted' object from the 'luz' package, this slot is defined as 'ANY' to accommodate various model structures.}

\item{\code{feature_names}}{A character vector specifying the names and order of the features used to train the model. This ensures that new data for prediction has the correct structure.}
}}

\author{
Weibin Huang <hwb2012@qq.com>
}
\keyword{classes}
