% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CCSR_DX_categories.R
\docType{data}
\name{CCSR_DX_categories}
\alias{CCSR_DX_categories}
\title{CCSR Diagnosis Categories}
\format{
A data frame with 547 obs. and 5 variables:
\describe{
  \item{chapter_abbr}{CCSR code prefix based on ICD-10 chapters}
  \item{chapter_number}{Numeric index for chapter (useful for plotting)}
  \item{chapter_desc}{Description of CCSR code chapter}
  \item{CCSR_category_code}{CCSR category code}
  \item{CCSR_category_desc}{Description of CCSR category}
}
}
\source{
\url{https://hcup-us.ahrq.gov/toolssoftware/ccsr/DXCCSR_v2023-1.zip}
}
\usage{
CCSR_DX_categories
}
\description{
A dataset containing the 2023-1 CCSR diagnosis categories.
}
\seealso{
\code{\link{CCSR_DX_mapping}} for ICD code mapping for each CCSR category.
}
\keyword{datasets}
