% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccs_base.R
\name{CHindex}
\alias{CHindex}
\title{Calinski-Harabasz index}
\usage{
CHindex(d = NULL, clustering, noisecluster = TRUE)
}
\arguments{
\item{d}{a distance object (as generated by \code{dist}) or a distance
    matrix between cases.}

\item{clustering}{an integer vector of length of the number of cases,
    which indicates a clustering. The clusters have to be numbered
    from 1 to the number of clusters.}

\item{noisecluster}{logical. If \code{TRUE}, it is assumed that the
    largest cluster number in \code{clustering} denotes a 'noise
    class', i.e. points that do not belong to any cluster. These points
    are not taken into account for the computation of all functions of
    within and between cluster distances including the validation
    indexes.}
}
\value{
Numeric. Calinski-Harabasz index.
}
\description{
Calinski-Harabasz index from fpc::cluster.stats
}
\seealso{
\code{\link[fpc]{cluster.stats}}; \code{\link{CCS-method.optimizeCluster}}; \code{\link{CCS-method.exploreCluster}}
}
\author{
Weibin Huang<\email{hwb2012@qq.com}>
}
