% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hyperTuningGS.R
\name{hyperTuningGS}
\alias{hyperTuningGS}
\title{Hyperparameter turning for GSClassifier models}
\usage{
hyperTuningGS(
  Xs_train,
  Ys_train,
  Xs_valid,
  Ys_valid,
  geneSet = NULL,
  geneAnnotation = NULL,
  geneid = "ensembl",
  para.grid = NULL,
  seed = 489,
  model.dir = "./hyperTuningGS/test01",
  verbose = TRUE,
  numCores = 2
)
}
\arguments{
\item{Xs_train}{Expression matrix of the training cohort}

\item{Ys_train}{Subtype vector of the training cohort}

\item{Xs_valid}{Expression matrix of the internal validation cohort}

\item{Ys_valid}{Subtype vector of the internal validation cohort}

\item{geneSet}{A list of genes for classification}

\item{geneAnnotation}{A data frame with ENSEMBL, SYMBOL and ENTREZID of the genes in \code{geneSet}}

\item{geneid}{type of gene id in \code{X}. One of \code{'symbol'}, \code{'entrez'} and \code{'ensembl'}.}

\item{para.grid}{a data frame of alternative parameters}

\item{model.dir}{the data of pr}

\item{verbose}{whether report messages}

\item{numCores}{No. of CPU core}
}
\value{
NULL. Data in \code{model.dir}
}
\description{
Hyperparameter turning for GSClassifier models
}
\details{
abc \cr
cde \cr
}
\examples{
## This is a simulative process and available only with CORRECT VARIABLES
}
\seealso{
\code{\link[GSClassifier]{parCallEnsemble}}; \code{\link[GSClassifier]{fitEnsembleModel}};
}
\author{
Weibin Huang<\email{hwb2012@qq.com}>
}
