% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scaller.R
\name{importance_dl}
\alias{importance_dl}
\title{Beta-version: Calculate Permutation Feature Importance for a luz Model}
\usage{
importance_dl(
  fitted_luz_model,
  feature_names,
  metric_name = "valid_loss",
  mode = "auto",
  plot = TRUE,
  ...
)
}
\arguments{
\item{fitted_luz_model}{A `luz_module_fitted` object, the result of `luz::fit()`.}

\item{feature_names}{A character vector of feature names. The order must match the columns of the input tensor `X`.}

\item{metric_name}{A string specifying the metric to monitor. This must be one of the names reported by `luz` in the validation metrics (e.g., "valid_loss", "valid_multiclass_auroc", "valid_accuracy").}

\item{mode}{A string, either "max" or "min" or "auto". This defines whether a higher value of the metric is better ("max") or a lower value is better ("min"). If "auto", the function infers the mode by checking if "loss" is in `metric_name`.}

\item{plot}{A logical value. If `TRUE` (the default), a ggplot of the feature importances is printed.}

\item{...}{Additional arguments (currently not used, for future extension).}
}
\value{
A `data.frame` containing the 'Feature' and its calculated 'Importance',
sorted from most to least important. If `plot = TRUE`, the data frame is
returned invisibly.
}
\description{
This function computes feature importance for a model trained with the 'luz' package using the permutation-based method. It measures the drop in model performance when a single feature's values are randomly shuffled.
}
\author{
Weibin Huang<\email{hwb2012@qq.com}>
}
