% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccs_base.R
\name{multiXClass_roc}
\alias{multiXClass_roc}
\title{Call ROC-AUC for Binary/multi-class response & multi-class predictor}
\usage{
multiXClass_roc(response, predictor, verbose = T)
}
\value{
a result from \link[pROC]{multiclass.roc}.
}
\description{
Call ROC-AUC for Binary/multi-class response & multi-class predictor
}
\author{
Weibin Huang<\email{hwb2012@qq.com}>
}
