% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phenotype.R
\name{newSubtype}
\alias{newSubtype}
\title{Get new normalized subtypes based on a record from \code{subtypePerformance}}
\usage{
newSubtype(x, record)
}
\arguments{
\item{x}{a character}

\item{record}{a record from \code{subtypePerformance}}
}
\value{
a character
}
\description{
Get new normalized subtypes based on a record from \code{subtypePerformance}
}
\seealso{
\code{\link{subtypePerformance }}.
}
\author{
Weibin Huang<\email{hwb2012@qq.com}>
}
