% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize.R
\name{normalize}
\alias{normalize}
\title{Normalization}
\usage{
normalize(object, .fun = CCS:::softmax)
}
\arguments{
\item{object}{\code{CCS} class or \code{d1}}

\item{.fun}{a function for normalization}
}
\value{
an object after normalization
}
\description{
Normalization for \code{CCS} class and \code{d1}
}
\examples{
## This is a simulative process and available only with CORRECT VARIABLES
}
\seealso{
\code{\link{ccs}}.
}
\author{
Weibin Huang<\email{hwb2012@qq.com}>
}
