% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phenotype.R
\name{plotSubtypeEffect}
\alias{plotSubtypeEffect}
\title{Plot for subtypeEffect}
\usage{
plotSubtypeEffect(data)
}
\arguments{
\item{data}{Data from \code{\link{subtypeEffect}} function}
}
\value{
Combined ggplot object
}
\description{
Plot for subtypeEffect
}
\author{
Weibin Huang</email{hwb2012@qq.com}>
}
