% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SelfssGSEA.R
\name{selfssGSEA}
\alias{selfssGSEA}
\title{selfssGSEA function}
\usage{
selfssGSEA(pathGenes = geneList, data)
}
\arguments{
\item{pathGenes}{Enter a vector,the self-help genes.Defult pathGenes is circadian clock gene list}

\item{data}{Enter your expression matrix,whose row is gene and column is sample}
}
\description{
selfssGSEA function
}
