% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulatePA.R
\name{simulatePA}
\alias{simulatePA}
\title{Simulate Probabilistic Array}
\usage{
simulatePA(
  nTissue = 15,
  nCohort = 10,
  nSample = 100,
  nCluster = 4,
  mean = 0,
  sd = 1,
  seed = 2023
)
}
\arguments{
\item{nTissue}{the amount of tissues}

\item{nCohort}{the amount of cohorts}

\item{nSample}{the amount of samples}

\item{nCluster}{the amount of clusters}

\item{mean}{the mean of Gaussian noise}

\item{sd}{the standard deviation of Gaussian noise}

\item{seed}{random seed}
}
\value{
A list of sample data, which includes the real types and related Probabilistic Array
}
\description{
Simulate Probabilistic Array
}
\examples{
## This is a simulative process and available only with CORRECT VARIABLES
}
\author{
Weibin Huang<\email{hwb2012@qq.com}>
}
