% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phenotype.R
\name{subtypeEffect}
\alias{subtypeEffect}
\title{Tumor clinical parameters based on molecular subtypes}
\usage{
subtypeEffect(
  data,
  col_subtype = "Subtype",
  col_cohort = "Cohort",
  col_response = "response",
  col_id = "SampleIDs",
  col_tissue = "tumor_type"
)
}
\arguments{
\item{data}{Data frame}

\item{col_subtype}{Colname of subtype}

\item{col_cohort}{Colname of cohort}

\item{col_response}{Colname of effect (positive=1, negative=0)}

\item{col_id}{Colname of sample ID}

\item{col_tissue}{Colname of source tissue}
}
\value{
Data frame containing useful clinical parameters:
\itemize{
  \item \code{RRMean} Mean response rate
  \item \code{RRMean_lower} Lower boundary of response rate
  \item \code{RRMean_upper} Upper boundary of response rate
  \item \code{NRRMean} Mean normalized response rate. The larger (absolute) the better.
  \item \code{NRRMean_lower} Lower boundary of normalized response rate
  \item \code{NRRMean_upper} Upper boundary of normalized response rate
  \item \code{H_index} Over \code{x}% cohorts with response rate over \code{x}%. The larger (raw) the better.
  \item \code{tissue_consistence_*} Consistency across cohorts in specified tissues. The larger (raw) the better.
}
}
\description{
Tumor clinical parameters based on molecular subtypes
}
\author{
Weibin Huang</email{hwb2012@qq.com}>
}
