% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phenotype.R
\name{subtypePerformance}
\alias{subtypePerformance}
\title{Downstream analysis: Molecular subtypes & Clinical features}
\usage{
subtypePerformance(
  data,
  col_subtype = "CCS",
  col_cohort = "Cohort",
  col_response = "response",
  col_id = "SampleIDs",
  col_tissue = "tissue",
  cohort_level = NULL,
  color_cohort = NULL,
  color_subtype = NULL,
  color_tissue = c(Brain = "#984EA3", Oral = "#BF5B17", Nasopharynx = "#8DD3C7", Lung =
    "#FF6868", Gastric = "#FF9843", Breast = "#A6D854", Colon = "#E31A1C", Kedney =
    "#86A7FC", Bladder = "#3468C0", Ovary = "#B3B3B3", Blood = "#F781BF", Skin =
    "#4DAF4A"),
  n_norm_subtype = 2,
  norm_roc_method = c("OptimizeRank", "Raw")[1],
  norm_rank_marker = c("NRR", "RR", "NRR+RR", "RR+NRR")[1],
  plot_ROCCutoff = 0.8,
  plot_layout_widths = c(9, 1),
  plot_silence_cohort = F,
  plot_silence_subtype = F,
  meta_method = c("DL", "REML", "ML", "PM", "SJ", "HE", "EB", "HK"),
  numCores = NULL,
  verbose = TRUE
)
}
\arguments{
\item{data}{Data frame}

\item{col_subtype}{Colname of subtype}

\item{col_cohort}{Colname of cohort}

\item{col_response}{Colname of effect (positive=1, negative=0)}

\item{col_id}{Colname of sample ID}

\item{col_tissue}{Colname of source tissue}

\item{cohort_level}{The level of cohort used in scater+bar plot}

\item{color_cohort}{Default is \code{NULL}, which means that all cohort names were colored as black. You can also use a named vector like \code{c('Kim2018_PD-1_gastric' = "#E31A1C")}.}

\item{color_subtype}{Default is \code{NULL}, which means that all raw subtype names were colored as \code{scales::hue_pal()(n)}. You can also use a named vector like \code{c('0' = "#E31A1C")}.}

\item{color_tissue}{a named vector like \code{c(Brain = "#984EA3",Oral = "#BF5B17")}}

\item{n_norm_subtype}{The number of normalized subtypes you want to explore}

\item{norm_roc_method}{The method to call ROC AUC. One of \code{'OptimizeRank'} and \code{'Raw'}.}

\item{norm_rank_marker}{The marker for normalization rank. Interpretation:
\itemize{
  \item \code{NRR} Ranked only by normalized response rate.
  \item \code{RR} Ranked only by raw response rate.
  \item \code{NRR+RR} Ranked by normalized response rate and then by raw response rate.
  \item \code{RR+NRR} Ranked by raw response rate and then by normalized response rate.
}}

\item{plot_ROCCutoff}{The cut-off value for visualization of ROC AUC}

\item{plot_layout_widths}{\code{layout_widths} for \code{\link[patchwork]{plot_layout}}}

\item{plot_silence_cohort}{Whether to silence cohort names in the plot}

\item{plot_silence_subtype}{Whether to silence subtype names in the plot}

\item{meta_method}{Random-effects estimator for the meta-analysis. One of \code{'DL'}, \code{'REML'}, \code{'ML'}, \code{'PM'} (Paule-Mandel), \code{'SJ'} (Sidik-Jonkman), \code{'HE'} (Hedges), \code{'EB'} (Empirical Bayes), or \code{'HK'} (Hartung-Knapp using REML + Knapp-Hartung adjustment).}

\item{numCores}{No. of CPU core}

\item{verbose}{whether report messages}
}
\value{
A list
\itemize{
  \item \code{Repeat} Data or parameters for repeatability
  \item \code{Plot} 1. Scatter plot + bar plot for RR & NRR; 2. Forest plot for RR.
  \item \code{Data} 1. ROC data; 2. Normalized subtype results 3. Clinic utility from \code{\link{subtypeEffect}}.
}
}
\description{
Downstream analysis: Molecular subtypes & Clinical features
}
\author{
Weibin Huang</email{hwb2012@qq.com}>
}
