% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{cluster,CCS-method}
\alias{cluster,CCS-method}
\title{CCS method: cluster}
\usage{
\S4method{cluster}{CCS}(object, method = "dbscan", distance = NULL, k = NULL, ...)
}
\arguments{
\item{object}{a \code{\link{CCS-class}} object}

\item{method}{Clustering methods. One of "\code{dbscan}", "\code{ward.D}", "\code{ward.D2}", "\code{single}", "\code{complete}", "\code{average}" (= UPGMA), "\code{mcquitty}" (= WPGMA), "\code{median}" (= WPGMC) or "\code{centroid}" (= UPGMC).}

\item{distance}{Distance method. Not work when \code{method="dbscan"}}

\item{k}{An integer scalar or vector with the desired number of groups.}

\item{...}{Other parameters for \code{\link[dbscan]{dbscan}} or \code{\link[stats]{hclust}}.}
}
\description{
\code{cluster} method for \code{CCS} class
}
\examples{
resCCS <- cluster(resCCS)
}
\seealso{
\code{\link{ccs}}.
}
