% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dr.R
\name{dr,CCS-method}
\alias{dr,CCS-method}
\title{CCS method: dr}
\usage{
\S4method{dr}{CCS}(
  object,
  method = c("UWOT", "UMAP", "t-SNE", "PCA")[1],
  dimension = c(2, 2),
  cover = F,
  model.dir = NULL,
  seed = 2024,
  verbose = T,
  ...
)
}
\arguments{
\item{object}{a \code{\link{CCS-class}} object}

\item{method}{method for the establishment of sub-models. One of \code{'GSClassifier'}}

\item{dimension}{the hyperparameters for 2 level t-SNE.}

\item{cover}{Whether to create an exact new dr result}

\item{model.dir}{a character. the path of model series.}

\item{seed}{random seed}

\item{verbose}{whether report messages}

\item{...}{Parameters of DR functions.}
}
\value{
data frame. Series of PC1, PC2 and etc after dimensionality reduction
}
\description{
\code{dr} method for \code{CCS} class. Dimensionality reduction (DR) for CCS probability matrix.
}
\details{
## Parameters

### Method == 't-SNE'

- \code{seed}: random seed
- \code{numCores}: CPU threads
}
\seealso{
\code{\link{ccs}}.
}
\author{
Weibin Huang<\email{hwb2012@qq.com}>
}
