% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{exploreCluster,CCS-method}
\alias{exploreCluster,CCS-method}
\title{CCS method: exploreCluster}
\usage{
\S4method{exploreCluster}{CCS}(object, nCluster = 2:20, model.dir = NULL, size = 15)
}
\description{
\code{exploreCluster} method for \code{CCS} class
}
\seealso{
\code{\link{ccs}}.
}
