% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{exploreCluster,CCS-method}
\alias{exploreCluster,CCS-method}
\title{CCS method: exploreCluster}
\usage{
\S4method{exploreCluster}{CCS}(
  object,
  nTop = 10,
  geom = c("cancer_type", "CCS"),
  hide.legend = c("cancer_type", "CCS")[2],
  model.dir = NULL,
  size = 15,
  width = 12,
  height = 9
)
}
\arguments{
\item{object}{a \code{\link{CCS-class}} object}

\item{nTop}{Integer. The count of top parameters you want to plot.}

\item{geom}{Some of \code{'cancer_type'} and \code{'CCS'}.}

\item{hide.legend}{Some of \code{'cancer_type'} and \code{'CCS'}. Which type of data should be hided in the plot legend.}

\item{model.dir}{a character. the path of model series.}

\item{size}{The plot size}

\item{width}{The plot width}

\item{height}{The plot height}
}
\description{
\code{exploreCluster} method for \code{CCS} class
}
\seealso{
\code{\link{ccs}}.
}
