% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importance.R
\name{importance,CCS-method}
\alias{importance,CCS-method}
\title{CCS method: importance}
\usage{
\S4method{importance}{CCS}(object, seed = 2024, verbose = TRUE, numCores = 16, ...)
}
\arguments{
\item{object}{a \code{\link{CCS-class}} object}

\item{seed}{random seed for \code{xgboost}}

\item{verbose}{whether report messages}

\item{numCores}{No. of CPU core}

\item{...}{parameters for \code{\link[xgboost]{xgb.cv}} and \code{\link[xgboost]{xgboost}}.}
}
\description{
\code{importance} method for \code{CCS} class
}
\examples{
To be continued!
}
\seealso{
\code{\link{ccs}};\code{\link{plotImportance}}.
}
