% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{optimizeCluster,CCS-method}
\alias{optimizeCluster,CCS-method}
\title{CCS method: optimizeCluster}
\usage{
\S4method{optimizeCluster}{CCS}(
  object,
  method = "dbscan",
  params = expand.grid(eps = c(0.02, 0.05, 0.1, seq(0.2, 1, 0.1)), minPts = c(5, 10, 15,
    20, 25, 30)),
  ...
)
}
\arguments{
\item{object}{a \code{\link{CCS-class}} object}

\item{...}{Other parameters for}
}
\description{
\code{optimizeCluster} method for \code{CCS} class
}
\examples{
resCCS <- optimizeCluster(resCCS)
}
\seealso{
\code{\link{ccs}}.
}
