% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccs.R
\name{plot.CCS}
\alias{plot.CCS}
\title{CCS method: plot}
\usage{
\method{plot}{CCS}(
  object,
  CCS = NULL,
  geom = c("cancer_type", "CCS"),
  hide.legend = c("cancer_type", "CCS")[2],
  rm.zero = TRUE,
  log.scale = FALSE,
  size = 15,
  verbose = TRUE
)
}
\arguments{
\item{object}{a \code{\link{CCS-class}} object}

\item{CCS}{Character. A vector of samples' CCS subtype.}

\item{geom}{Some of \code{'cancer_type'} and \code{'CCS'}.}

\item{hide.legend}{Some of \code{'cancer_type'} and \code{'CCS'}. Which type of data should be hided in the plot legend.}

\item{rm.zero}{Whether to remove zero value of CCS subtypes. In the strategy of \code{dbscan}, 0 means uncategorized, so you should set \code{rm.zero = TRUE}(default).}

\item{log.scale}{Whether to show x/y axis in the log scale.}

\item{size}{The plot size}

\item{verbose}{whether report messages}
}
\value{
plot: A ggplot2 object
}
\description{
\code{plot} method for \code{CCS} class. Plot dimension scatter plot for CCS results.
}
\examples{
ccs_plot <- plot(resCCS)
}
\seealso{
\code{\link{ccs}}.
}
\author{
Weibin Huang<\email{hwb2012@qq.com}>
}
