% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccs.R
\name{plot.CCS}
\alias{plot.CCS}
\title{CCS method: plot}
\usage{
\method{plot}{CCS}(object, CCS = NULL, geom = c("cancer_type", "CCS")[2], size = 15)
}
\arguments{
\item{object}{a CCS object}

\item{CCS}{Character. A vector of samples' CCS subtype.}

\item{geom}{Some of \code{'cancer_type'} and \code{'CCS'}.}

\item{size}{the size of ggplot}
}
\value{
plot: A ggplot2 object
}
\description{
\code{plot} method for \code{CCS} class
}
\examples{
ccs_plot <- plot(resCCS)
}
\seealso{
\code{\link{ccs}}.
}
\author{
Weibin Huang<\email{hwb2012@qq.com}>
}
