% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccs.R
\name{plotCancerSubtype,CCS-method}
\alias{plotCancerSubtype,CCS-method}
\title{CCS method: plotCancerSubtype}
\usage{
\S4method{plotCancerSubtype}{CCS}(
  object,
  CCS = NULL,
  rm.zero = TRUE,
  add.cell_fun = TRUE,
  size = 15,
  col = NULL,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{object}{a \code{\link{CCS-class}} object}

\item{rm.zero}{Whether to remove zero value of CCS subtypes. In the strategy of \code{dbscan}, 0 means uncategorized, so you should set \code{rm.zero = TRUE}(default).}

\item{add.cell_fun}{Logic. Whether add a cell_fun into the heatmap.}

\item{size}{The plot size}

\item{col}{A vector of colors if the color mapping is discrete or a color mapping  function if the matrix is continuous numbers (should be generated by \code{\link[circlize]{colorRamp2}}). If the matrix is continuous, the value can also be a vector of colors so that colors can be interpolated. Pass to \code{\link[ComplexHeatmap]{ColorMapping}}. For more details and examples, please refer to \url{https://jokergoo.github.io/ComplexHeatmap-reference/book/a-single-heatmap.html#colors} .}

\item{verbose}{whether report messages}

\item{...}{Arguments to function\code{\link[ComplexHeatmap]{Heatmap}}.}
}
\value{
a complete \code{\link[ComplexHeatmap]{Heatmap}} class.
}
\description{
\code{plotCancerSubtype} method for \code{CCS} class. Plot CCS subtype across cancer types.
}
