% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importance.R
\name{plotImportance,CCS-method}
\alias{plotImportance,CCS-method}
\title{CCS method: plotImportance}
\usage{
\S4method{plotImportance}{CCS}(
  object,
  type = c("tissue", "cohort", "feature")[1],
  convertGene = F,
  nTop = 10,
  size = 10,
  ...
)
}
\arguments{
\item{object}{a \code{\link{CCS-class}} object}

\item{type}{One of \code{"tissue"}, \code{"cohort"}, or \code{"feature"}.}

\item{convertGene}{Whether to convert names like \code{"ENSG00000185811:ENSG00000187189"} to names like \code{"IKZF1:TSPYL4"}.}

\item{size}{The plot size}

\item{...}{parameters to \code{\link[ggplot2]{geom_bar}}.}
}
\description{
\code{plotImportance} method for \code{CCS} class
}
\examples{
To be continued!
}
\seealso{
\code{\link{ccs}}; \code{\link{importance}}.
}
