% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccs.R
\name{predict.CCS}
\alias{predict.CCS}
\title{CCS method: predict}
\usage{
\method{predict}{CCS}(object, X, model.dir = NULL, verbose = T, numCores = 4)
}
\arguments{
\item{object}{a CCS object}

\item{X}{a numeric RNA expression vector or matrix with gene names}

\item{model.dir}{a character. the path of model series.}

\item{verbose}{whether report messages}

\item{numCores}{No. of CPU core}
}
\value{
predict: A list object with CCS subtype prediction.
}
\description{
\code{predict} method for \code{CCS} class
}
\examples{
ccs_pred <- predict(resCCS, model.dir = "./ccs/project_01")
}
\seealso{
\code{\link{ccs}}.
}
\author{
Weibin Huang<\email{hwb2012@qq.com}>
}
