% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scaller.R
\name{scaller,CCS-method}
\alias{scaller,CCS-method}
\title{CCS method: scaller}
\usage{
\S4method{scaller}{CCS}(
  object,
  model.dir = NULL,
  Prop = 0.8,
  seed = 456,
  cover = FALSE,
  numCores = 6,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{object}{a \code{\link{CCS-class}} object}

\item{model.dir}{a character. the path of model series.}

\item{Prop}{the proportion of training cohort}

\item{seed}{random seed}

\item{cover}{Whether to cover the existed result}

\item{numCores}{No. of CPU core}

\item{verbose}{whether report messages}

\item{...}{Parameters for \code{\link[xgboost]{xgboost}}.}
}
\description{
\code{scaller} method for \code{CCS} class
}
\details{
The CCS object must be modified by \code{\link{CCS-method.cluster}} or \code{\link{CCS-method.optimizeCluster}}.
}
\examples{
resCCS <- scaller(resCCS)
}
\seealso{
\code{\link{ccs}};\code{\link[xgboost]{xgboost}};
}
