% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccs.R
\name{ccs}
\alias{ccs}
\title{Cohort Congress System}
\usage{
ccs(
  data,
  geneSet,
  geneAnnotation,
  geneid = "ensembl",
  params = list(nfold = 5, nrounds = 100, nthread = 2, eta = 0.5, gamma = 0, max_depth =
    10, colsample_bytree = 1, min_child_weight = 1, subsample = 0.7, n = 100, sampSize =
    0.7, ptail = 0.2),
  seed = 489,
  min.nc = 2,
  max.nc = 10,
  model.dir = "./ccs/project_01",
  verbose = T,
  numCores = 4
)
}
\arguments{
\item{data}{a list with components (expression matrix + subtype vector)}

\item{geneSet}{A list of genes for classification}

\item{geneAnnotation}{A data frame with ENSEMBL, SYMBOL and ENTREZID of the genes in \code{geneSet}}

\item{geneid}{type of gene id in \code{X}. One of \code{'symbol'}, \code{'entrez'} and \code{'ensembl'}.}

\item{params}{parameters like X function}

\item{model.dir}{a character. the path of model series.}

\item{verbose}{whether report messages}

\item{numCores}{No. of CPU core}
}
\value{
a CCS object
}
\description{
Cohort Congress System
}
\examples{
## This is a simulative process and available only with CORRECT VARIABLES
}
\author{
Weibin Huang<\email{hwb2012@qq.com}>
}
