% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccs.R
\name{ccs}
\alias{ccs}
\title{Cohort Congress System}
\usage{
ccs(
  data,
  geneSet,
  geneAnnotation,
  geneid = "ensembl",
  params = list(nfold = 5, nrounds = 100, nthread = 2, eta = 0.5, gamma = 0, max_depth =
    10, colsample_bytree = 1, min_child_weight = 1, subsample = 0.7, n = 100, sampSize =
    0.7, ptail = 0.2),
  seed = 489,
  dimension = c(2, 2),
  perplexity = 30,
  theta = 0.3,
  random.cluster = TRUE,
  params.NbClust = list(distance = "euclidean", min.nc = 2, max.nc = 10, method =
    "ward.D2", index = "all"),
  model.dir = "./ccs/project_01",
  verbose = TRUE,
  numCores = 4
)
}
\arguments{
\item{data}{a list with components (expression matrix + subtype vector)}

\item{geneSet}{A list of genes for classification}

\item{geneAnnotation}{A data frame with ENSEMBL, SYMBOL and ENTREZID of the genes in \code{geneSet}}

\item{geneid}{type of gene id in \code{X}. One of \code{'symbol'}, \code{'entrez'} and \code{'ensembl'}.}

\item{params}{parameters of \code{\link[xgboost]{xgb.cv}}, \code{\link[xgboost]{xgboost}}, and \code{\link[GSClassifier]{fitEnsembleModel}}.}

\item{seed}{Seeds for \code{\link[Rtsne]{Rtsne}} function.}

\item{dimension}{the hyperparameters for 2 level t-SNE.}

\item{perplexity}{numeric; Perplexity parameter (should not be bigger than 3 * perplexity < nrow(X) - 1, see details for interpretation)}

\item{theta}{numeric; Speed/accuracy trade-off (increase for less accuracy), set to 0.0 for exact TSNE (default: 0.5)}

\item{random.cluster}{Whether to use random samples to decide the best subtype number. Default is \code{TRUE}, which is faster but more vague comparing with \code{FALSE} when \code{data} contains lots of samples (>10000, for example).}

\item{params.NbClust}{Parameters of \code{CCS:::NbClust2}. Possible options are:
\itemize{
  \item \code{distance} the distance measure to be used to compute the dissimilarity matrix. This must be one of: "euclidean", "maximum", "manhattan", "canberra", "binary", "minkowski" or "NULL". By default, distance="euclidean". If the distance is "NULL", the dissimilarity matrix (diss) should be given by the user. If distance is not "NULL", the dissimilarity matrix should be "NULL".
  \item \code{method} the cluster analysis method to be used. This should be one of: "ward.D", "ward.D2", "single", "complete", "average", "mcquitty", "median", "centroid", "kmeans".
  \item \code{index} the index to be calculated. This should be one of : "kl", "ch", "hartigan", "ccc", "scott", "marriot", "trcovw", "tracew", "friedman", "rubin", "cindex", "db", "silhouette", "duda", "pseudot2", "beale", "ratkowsky", "ball", "ptbiserial", "gap", "frey", "mcclain", "gamma", "gplus", "tau", "dunn", "hubert", "sdindex", "dindex", "sdbw", "all" (all indices except GAP, Gamma, Gplus and Tau), "alllong" (all indices with Gap, Gamma, Gplus and Tau included). Cory's recommendation: one of "ch", "duda", "cindex", "gamma", "beale", "gap", and "all".
 \item \code{min.nc} minimal number of clusters, between 1 and (number of objects - 1)
}}

\item{model.dir}{a character. the path of model series.}

\item{verbose}{whether report messages}

\item{numCores}{No. of CPU core}
}
\value{
a CCS object
}
\description{
Cohort Congress System
}
\examples{
## This is a simulative process and available only with CORRECT VARIABLES
}
\references{
\href{https://cran.r-project.org/web/packages/NbClust/index.html}{1. NbCluster package} \cr
\href{http://www.sthda.com/english/wiki/wiki.php?id_contents=7940}{2. A tutorial - DBSCAN: density-based clustering for discovering clusters in large datasets with noise}
}
\seealso{
\code{\link[xgboost]{xgb.cv}}; \code{\link[xgboost]{xgboost}}; \code{\link[NbClust]{NbClust}};
}
\author{
Weibin Huang<\email{hwb2012@qq.com}>
}
