% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccs.R
\name{ccs}
\alias{ccs}
\title{Cohort Congress System}
\usage{
ccs(
  data,
  geneSet,
  geneAnnotation,
  method = "GSClassifier",
  geneid = "ensembl",
  params = list(nfold = 5, nrounds = 100, nthread = 2, eta = 0.5, gamma = 0, max_depth =
    10, colsample_bytree = 1, min_child_weight = 1, subsample = 0.7, n = 100, sampSize =
    0.7, ptail = 0.2),
  seed = 489,
  model.dir = "./ccs/project_01",
  verbose = TRUE,
  numCores = 4
)
}
\arguments{
\item{data}{a list with components (expression matrix + subtype vector)}

\item{geneSet}{A list of genes for classification}

\item{geneAnnotation}{A data frame with ENSEMBL, SYMBOL and ENTREZID of the genes in \code{geneSet}}

\item{method}{method for the establishment of sub-models. One of \code{'GSClassifier'}}

\item{geneid}{type of gene id in \code{X}. One of \code{'symbol'}, \code{'entrez'} and \code{'ensembl'}.}

\item{params}{parameters of \code{\link[xgboost]{xgb.cv}}, \code{\link[xgboost]{xgboost}}, and \code{\link[GSClassifier]{fitEnsembleModel}}. Some important options are:
\itemize{
  \item \code{nfold} No. of cross validation subcohorts.
  \item \code{nrounds} Max number of boosting iterations. GSClassifier have optimized nrounds, so here I set a large value.
  \item \code{nthread} No. of CPU cores used in \code{\link[xgboost]{xgboost}}.
  \item \code{eta} Step size shrinkage used in update to prevent overfitting. Range = [0,1]
  \item \code{gamma} The larger gamma is, the more conservative the algorithm will be. Range = [0,∞]
  \item \code{max_depth} Maximum depth of a tree. Deeper trees can capture more complex patterns in the data, but may also lead to overfitting. Range = [0,∞]
  \item \code{colsample_bytree} Percentage of columns used for each tree construction. Lowering this value can prevent overfitting by training on a subset of the features. Range = (0, 1]
  \item \code{min_child_weight} The larger min_child_weight is, the more conservative the algorithm will be. Range = [0,∞]
  \item \code{subsample} Preventing overfitting
  \item \code{n} \code{\link[GSClassifier]{fitEnsembleModel}} parameter
  \item \code{sampSize} \code{\link[GSClassifier]{fitEnsembleModel}} parameter. Range = (0,1]
  \item \code{ptail} \code{\link[GSClassifier]{fitEnsembleModel}} parameter. Range = (0,0.5]
}}

\item{model.dir}{a character. the path of model series.}

\item{verbose}{whether report messages}

\item{numCores}{No. of CPU core}
}
\value{
a CCS object
}
\description{
Cohort Congress System
}
\examples{
## This is a simulative process and available only with CORRECT VARIABLES
}
\references{
\href{https://cran.r-project.org/web/packages/NbClust/index.html}{1. NbCluster package} \cr
\href{http://www.sthda.com/english/wiki/wiki.php?id_contents=7940}{2. A tutorial - DBSCAN: density-based clustering for discovering clusters in large datasets with noise} \cr
\href{https://satijalab.org/seurat/articles/get_started_v5_new}{Getting Started with Seurat • Seurat}
}
\seealso{
\code{\link[xgboost]{xgb.cv}}; \code{\link[xgboost]{xgboost}}; \code{\link[NbClust]{NbClust}};
}
\author{
Weibin Huang<\email{hwb2012@qq.com}>
}
