% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phenotype.R
\name{subtypePerformance}
\alias{subtypePerformance}
\title{Downstream analysis: Molecular subtypes & Clinical features}
\usage{
subtypePerformance(
  data,
  col_subtype = "CCS",
  col_cohort = "Cohort",
  col_response = "response",
  col_id = "SampleIDs",
  col_tissue = "tissue",
  cohort_level = NULL,
  color_tissue = c(Brain = "#984EA3", Oral = "#BF5B17", Nasopharynx = "#8DD3C7", Lung =
    "#FF6868", Gastric = "#FF9843", Breast = "#A6D854", Colon = "#E31A1C", Kedney =
    "#86A7FC", Bladder = "#3468C0", Ovary = "#B3B3B3", Blood = "#F781BF", Skin =
    "#4DAF4A"),
  n_norm_subtype = 2,
  plot_layout_widths = c(9, 1),
  numCores = NULL,
  verbose = TRUE
)
}
\arguments{
\item{data}{Data frame}

\item{col_subtype}{Colname of subtype}

\item{col_cohort}{Colname of cohort}

\item{col_response}{Colname of effect (positive=1, negative=0)}

\item{col_id}{Colname of sample ID}

\item{col_tissue}{Colname of source tissue}

\item{cohort_level}{The level of cohort used in scater+bar plot}

\item{color_tissue}{a named vector like \code{c(Brain = "#984EA3",Oral = "#BF5B17")}}

\item{n_norm_subtype}{The number of normalized subtypes you want to explore}

\item{plot_layout_widths}{\code{layout_widths} for \code{\link[patchwork]{plot_layout}}}

\item{numCores}{No. of CPU core}

\item{verbose}{whether report messages}
}
\value{
A list
\itemize{
  \item \code{Repeat} Data or parameters for repeatability
  \item \code{Plot} 1. Scatter plot + bar plot for RR & NRR; 2. Forest plot for RR.
  \item \code{Data} 1. ROC data; 2. Normalized subtype results 3. Clinic utility from \code{\link{subtypeEffect}}.
}
}
\description{
Downstream analysis: Molecular subtypes & Clinical features
}
\author{
Weibin Huang</email{hwb2012@qq.com}>
}
